Instantiate a new minibuffer on every call and add a status bar.
The status bar is (for now) still a minibuffer, albeit one that's only
instantiated once per window.

To allow nesting of minibuffers, we add a `active-minibuffers' stack to the
window.
When a minibuffer is popped, the previous minibuffer is restores thanks to the
HTML content which was saved in the `content' slot.