Remove most INTERFACE function arguments.
This greatly simplifies the code.
To deal with multiple interfaces, it's enough to let-bind `*interface*'.

Most of the RPC functions used to be methods on REMOTE-INTERFACE, so I've turned
them all into functions.  The "defmethod" was only used for typing, not for
method dispatching; thus I've typed all functions so that we don't lose the
typing information in the process.

The ACTIVE-BUFFER defun would clash with the WINDOW accessor of the same name, so I
had to rename it to CURRENT-BUFFER.
For consistency, I've renamed the MINIBUFFER function to CURRENT-MINIBUFFER.